runner: handle timeout in client
Ducktape sets up a server and uses multiprocessing to spawn a client.
They communicate via TCP sockets, with the client sending events about
the state of the running test. The --test-runner-timeout flag is an
integer that defines the maximum time the client has to complete the
test. If the server doesn't receive any event from the client within
that time, it raises an exception (server-side), terminates all
processes, and exits with code 1.

This behavior makes it difficult to diagnose test e.g. where the
test was hanging before the socket timeout occurred. Additionally,
such tests are not included in the generated reports.

This change moves the timeout logic to the client. The runner timeout is
now passed to the client, which raises a TestTimeoutError if the test
doesn't finish in time. It uses Linux signals to enforce the timeout,
similar to how pytest-timeout handles test timeouts:
https://github.com/pytest-dev/pytest-timeout/blob/2e96621ee1b61057438c0f8c5b158eddc31654c5/pytest_timeout.py#L304

If a timeout occurs, the client raises an exception just like a regular
assertion error. This allows the test to be reported properly and avoids
sudden termination of the test suite. Using signals also provides a full
traceback, making it clear where the test was hanging when the timeout
was triggered.

jira: [DEVPROD-3200]