handle internal exceptions in _do_run
Exceptions throw in the test code are well handled in ducktape, but it
also possible for the surrounding code in runner_client to throw
exceptions itself. E.g., because of failure to communicate with the
parent, or because an exception class (from a test) __str__ method
throws, or other issues.

In this case, the test summary ends up as "Test passed", though the
test is marked as failed, which is very confusing. Update the logic
to detect this case and return a summary indicating an internal
exception occurred, as well as the str() of the exception.