runner: more details when a child fails to stop
When a child fails to stop, i.e., it finishes teardown but then does
not exit in a timely manner, we log a message about it. This change
updates the message to indicate clearly which test is involved, and
also to include language indicating this is a "corrupting" failure,
i.e., one that may cause other tests to fail later on.

This is very likely for this failure since the runner will be busy for
30s waiting for the child to exit, during which time it cannot service
messages from other concurrent tests on the message queue, and these
have timeouts as short as 15s (e.g., for a log message) after which
they will fail.

Include also the string "CORRUPTING_FAILURE" in the log message as a
tag we can search for which could include other corrupting failures
in the future.